+++ src/main/java/com/clubber/ClubberServer/global/error/GlobalExceptionHandler.java
@@ -1,13 +1,8 @@
 package com.clubber.ClubberServer.global.error;
 
-import java.io.PrintWriter;
-import java.io.StringWriter;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 
-import com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.client.discord.DiscordClient;
-import com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.dto.discord.DiscordMessage;
+import com.clubber.ClubberServer.global.event.exceptionalarm.ExceptionAlarmEventPublisher;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -33,7 +28,7 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
 
-	public final DiscordClient discorClient;
+	public final ExceptionAlarmEventPublisher publisher;
 
 	@ExceptionHandler(BaseException.class)
 	public ResponseEntity<ErrorResponse> handleBaseException(
@@ -84,47 +79,13 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
 		ErrorResponse errorResponse = new ErrorResponse(status.value(), errorMessages.toString(), uri);
 		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
 	}
-    
+
 	@ExceptionHandler({Exception.class})
 	public ResponseEntity<Object> handleAllException(Exception e, WebRequest request) {
 		GlobalErrorCode internalServerError = GlobalErrorCode.INTERNAL_SERVER_ERROR;
 		log.error("INTERNAL SERVER ERROR", e);
-		sendDiscordAlarm(e, request);
+		publisher.throwExceptionAlarmEvent(e, request);
 		return ResponseEntity.status(internalServerError.getStatus())
 			.body(internalServerError.getErrorReason());
-	}
-
-	private void sendDiscordAlarm(Exception e, WebRequest request) {
-		discorClient.sendAlarm(createDiscordMessage(e, request));
-	}
-
-	private DiscordMessage createDiscordMessage(Exception e, WebRequest request) {
-		List<DiscordMessage.Embed> embedList = List.of(DiscordMessage.Embed
-				.builder()
-				.title("에러 테스트중입니다.")
-				.description(getDescription(e, request))
-				.build());
-
-		return DiscordMessage.builder()
-				.content("에러 발생 내용")
-				.embeds(embedList)
-				.build();
-	}
-
-	private String getErrorStackTrace(Exception e) {
-		StringWriter stringWriter = new StringWriter();
-		e.printStackTrace(new PrintWriter(stringWriter));
-		return stringWriter.toString();
-	}
-
-	private String getRequestFullPath(WebRequest webRequest) {
-		HttpServletRequest request = ((ServletWebRequest) webRequest).getRequest();
-		return request.getMethod() + request.getRequestURL();
-	}
-
-	private String getDescription(Exception e, WebRequest request) {
-		return "발생 시간 : " + LocalDateTime.now() + "\n"
-				+ "요청 URL : " + getRequestFullPath(request) + "\n"
-				+ "에러 사항 : " + getErrorStackTrace(e).substring(0, 1000) + "\n";
 	}
 }
