+++ src/main/java/com/clubber/ClubberServer/global/event/exceptionalarm/ExceptionAlarmEvent.java
@@ -1,0 +1,12 @@
+package com.clubber.ClubberServer.global.event.exceptionalarm;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.web.context.request.WebRequest;
+
+@Getter
+@AllArgsConstructor
+public class ExceptionAlarmEvent {
+    private Exception e;
+    private WebRequest request;
+}
